requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.57.1
diffusers==0.35.2
datasets==4.4.1
accelerate==1.11.0
safetensors==0.6.2
tensorflow==2.19.1
tf_keras==2.20.1
keras==3.11.2
opencv-python==4.12.0.88
pillow==11.3.0
scikit-image==0.25.2
mtcnn==1.0.0
retina-face==0.0.17
deepface==0.0.95
numpy==2.1.3
pandas==2.3.3
scipy==1.16.3
einops==0.8.1
tqdm==4.67.1
gdown==5.2.0
huggingface-hub==0.36.0
flask==3.1.2
flask-cors==6.0.1
requests==2.32.5
aiohttp==3.13.2